reformating all blades, adding header footer and blacknav for common usages
