--- Domashna 4/healthe_app/public/css/results.css
@@ -1,3 +1,0 @@
-body {
-    background-color: #24526d;
-}
