--- Domashna 4/healthe_app/resources/js/app.js
@@ -1,1 +1,0 @@
-require('./bootstrap');
