+++ Domashna 4/healthe_app/resources/views/common/blackNav.blade.php
@@ -1,0 +1,15 @@
+<nav class="navbar navbar-expand-sm navbar-dark bg-dark" id="blackNav">
+    <div class="collapse navbar-collapse justify-content-between">
+        <ul class="navbar-nav mr-auto">
+            <li class="nav-item">
+                <a class="nav-link" href="{{ URL::route('home') }}">Почетна</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="{{ URL::route('about') }}">За нас</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="{{ URL::route('help') }}">Помош</a>
+            </li>
+        </ul>
+    </div>
+</nav>
