+++ Domashna 4/healthe_app/resources/views/common/footer.blade.php
@@ -1,0 +1,6 @@
+<!--MapBox JS -->
+<script src="{{ asset('js/mapbox.js') }}"></script>
+<!-- Bootstrap -->
+<script src="{{ asset('js/app.js') }}"></script>
+</body>
+</html>
